Refactor: UI / Update css of widget components #605
