+++ client/src/components/Asset/Status/AssetCardStatus.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { AssetProps } from '../../../screens/Asset';
-import Carousel from '../../UI/Carousel';
+import Carousel from '../../UI/widget/Carousel';
 import AssetCardItemEditor from '../Editor/AssetCardItemEditor';
 import AssetCardListEditor from '../Editor/AssetCardListEditor';
 import classes from './AssetCardStatus.module.css';
