--- client/src/components/UI/Calendar.module.css
@@ -1,60 +1,0 @@
-.container h6 {
-    text-align: center;
-}
-
-.container table {
-    text-align: center;
-    table-layout: fixed;
-    width: 100%;
-    height: 100%;
-}
-
-.container tbody {
-    overflow-y: scroll;
-}
-
-.day {
-    font-size: var(--size-9);
-}
-
-.month {
-    height: 0;
-    font-size: var(--size-7);
-    font-weight: 600;
-}
-
-.date {
-    position: relative;
-    padding-top: var(--size-11);
-    font-size: var(--size-7);
-    vertical-align: top;
-    font-weight: 300;
-}
-
-.date:hover {
-    font-weight: 600;
-    cursor: pointer;
-    background-color: #f6f6f6;
-    border-radius: 0.5rem;
-}
-
-.today {
-    font-weight: 800;
-    color: var(--primary);
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-.today::after {
-    content: '';
-    display: block;
-    background-color: var(--primary);
-    width: 0.25rem;
-    height: 0.25rem;
-    border-radius: 50%;
-}
-
-.after {
-    opacity: 0.45;
-}
