--- client/src/components/UI/Carousel.module.css
@@ -1,52 +1,0 @@
-.container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-.carouselList {
-    display: flex;
-    width: 100%;
-    overflow-x: scroll;
-    -ms-overflow-style: none;
-    scrollbar-width: 0;
-    scroll-snap-type: x mandatory;
-    scroll-behavior: smooth;
-}
-
-.carouselList::-webkit-scrollbar {
-    display: none;
-}
-
-.carouselItem {
-    display: flex;
-    flex-shrink: 0;
-    flex-direction: column;
-    justify-content: space-around;
-    align-items: center;
-    width: 100%;
-    scroll-snap-align: start;
-    scroll-snap-stop: always;
-}
-
-.carouselNav {
-    display: flex;
-    justify-content: space-around;
-}
-
-.carouselNav button {
-    margin: 0 var(--size-9);
-}
-
-.carouselNav button {
-    width: var(--size-10);
-    height: var(--size-10);
-    border: none;
-    border-radius: 50%;
-    background-color: #e0e0e0;
-}
-
-.carouselNav .active {
-    cursor: inherit;
-    background-color: #0e0e0e;
-}
