+++ client/src/css/color.css
@@ -73,14 +73,6 @@
   background-color: var(--gray-100);
 }
 
-.bg-primary {
-  background-color: var(--primary);
-}
-
-.bg-secondary {
-  background-color: var(--secondary);
-}
-
 .bg-white {
   background-color: var(--white);
 }
@@ -95,6 +87,14 @@
 
 .bg-gray-0 {
   background-color: var(--gray-0);
+}
+
+.bg-primary {
+  background-color: var(--primary);
+}
+
+.bg-secondary {
+  background-color: var(--secondary);
 }
 
 /* fill */
