+++ client/src/screens/Budget.tsx
@@ -18,7 +18,7 @@ import TransactionForm from '../components/Budget/Transaction/TransactionForm';
 import TransactionList from '../components/Budget/Transaction/TransactionList';
 import TransactionNav from '../components/Budget/Transaction/TransactionNav';
 import BudgetHeader from '../components/Budget/UI/BudgetHeader';
-import Carousel from '../components/UI/Carousel';
+import Carousel from '../components/UI/widget/Carousel';
 import EmojiOverlay from '../components/UI/overlay/EmojiOverlay';
 import DefaultStatus from '../components/User/Default/DefaultStatus';
 import { useAppDispatch } from '../hooks/useRedux';
