+++ client/src/screens/NewBudget.tsx
@@ -5,7 +5,7 @@ import BudgetList from '../components/Budget/List/BudgetList';
 import TotalStatus from '../components/Budget/Status/TotalStatus';
 import BudgetHeader from '../components/Budget/UI/BudgetHeader';
 import Button from '../components/UI/button/Button';
-import Carousel from '../components/UI/Carousel';
+import Carousel from '../components/UI/widget/Carousel';
 import { useAppDispatch, useAppSelector } from '../hooks/useRedux';
 import { uiActions } from '../store/ui';
 import { createBudgetFromBasic } from '../util/api/budgetAPI';
